addon manager, coredns
still seeing issues with cilium on node restarts, the operator is
getting unauthorized when attempting to connect to the api server.
Running a rollout restart resolves the issue. I wonder if our service
account tokens are revoked due to the service account cert and key being
re-issued on kubelet restarts.